--- model/src/main/java/uk/gov/hmcts/payment/api/dto/CreditAccountStatusHistoryDto.java
@@ -1,32 +1,0 @@
-package uk.gov.hmcts.payment.api.dto;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.databind.PropertyNamingStrategy;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import lombok.*;
-
-import java.util.Date;
-
-import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
-
-@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
-@JsonInclude(NON_NULL)
-@Builder(builderMethodName = "creditAccountStatusHistoryDtoWith")
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class CreditAccountStatusHistoryDto {
-    private String status;
-    private String externalStatus;
-
-    private String errorCode;
-    private String errorMessage;
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSZ", timezone = "GMT")
-    private Date dateCreated;
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSZ", timezone = "GMT")
-    private Date dateUpdated;
-}
